+++ R/numericalDifferentiation.R
@@ -4,15 +4,17 @@
 #'   of the given function at the given point using central difference numerical approximation.
 #'   \code{get_gradient} or \code{get_hessian} calculates the gradient or Hessian matrix of the
 #'   log-likelihood function at the parameter estimates of class \code{'gmvar'} object. \code{get_soc}
-#'   returns eigenvalues of the Hessian matrix.
+#'   returns eigenvalues of the Hessian matrix, and \code{get_foc} is the same as \code{get_gradient}
+#'   but named conveniently.
 #'
 #' @inheritParams quantile_residuals
 #' @param x a numeric vector specifying the point where the gradient or Hessian should be calculated.
 #' @param fn a function that takes in argument \code{x} as the \strong{first} argument.
 #' @param h difference used to approximate the derivatives.
 #' @param ... other arguments passed to \code{fn}
-#' @details Especially the functions \code{get_gradient} or \code{get_hessian} can be used to check whether
-#'   the found estimates denote a (local) maximum point, a saddle point or something else.
+#' @details Especially the functions \code{get_foc} or \code{get_soc} can be used to check whether
+#'   the found estimates denote a (local) maximum point, a saddle point or something else. Note that
+#'   profile log-likelihoods can be conveniently plotted with the function \code{profile_logliks}.
 #' @return Gradient functions return numerical approximation of the gradient, and Hessian functions return
 #'   numerical approximation of the Hessian. \code{get_soc} returns eigenvalues of the Hessian matrix.
 #' @seealso \code{\link{profile_logliks}}
@@ -101,4 +103,10 @@ get_hessian <- function(gmvar, h=6e-06) {
 #' @export
 get_soc <- function(gmvar, h=6e-6) {
   eigen(get_hessian(gmvar, h))$value
+}
+
+#' @rdname calc_gradient
+#' @export
+get_foc <- function(gmvar, h=6e-6) {
+  get_gradient(gmvar, h=h)
 }