+++ NEWS.md
@@ -18,5 +18,9 @@
 
 # gmvarkit 1.1.2
 
-* Improvements on the comments and documentation.
+* New functions: 'profile_logliks' (plot profile log-likelihood functions), 'get_foc' (gradient at the estimates, remember also 'get_soc')
+* Now standard errors are printed correctly for models imposing all kinds of constraints (in earlier versions standard errors for
+  such constrained AR parameters that involved sums or multiplication were incorrect).
 * Minor update on the genetic algorithm.
+* Minor update on the print and summary-print methods. 
+* Improvements on the comments and documentation.